Sample blinking LED code
I finally got some results in using my Arduino to program the stand-alone ATmega328 with no Arduino bootloader!!! The code below is C code that I just pasted into the Arduino IDE, and it understands it! The only issue is that the timing seems off.
If I set the timer to 100ms, for example, it actually blinks about once every 1.5 seconds. I suspect it has something to do with 
defining the clock speed of the chip, but I am not sure.